+++ monitor/model/models/model.go
@@ -11,14 +11,14 @@ type Params struct {
 
 // Model interface defined of models
 type Model interface {
-	Update(param uint8, data interface{}) error
+	Update(param uint32, data interface{}) error
 	Save(path string) error
 	This() interface{}
 	GetParams() *Params
 	GetData() 	interface{}
 }
 
-// Check check if model match the wanted model
+// Check if model match the wanted model
 func (params *Params) Check(modelType enum.ModelType, instanceId uint32) bool {
 	return params.ModelType == modelType && params.InstanceID == instanceId
 }
